Feat: Update PK type `Long` -> `Integer`

+++ Backend/src/main/java/com/solmi/shorket/market/domain/MarketImg.java
@@ -17,7 +17,7 @@ public class MarketImg {
     @Id
     @GeneratedValue
     @Column(name = "store_img_idx")
-    private Long idx;
+    private Integer idx;
 
     @NotNull
     @ManyToOne
